requirements.txt
pyobjc-framework-Quartz
pyautogui
screeninfo